+++ platform/swift/source/CaptureRustBridge.h
@@ -35,6 +35,7 @@ void capture_report_error(const char *message);
  * @param network the Capture Network protocol to use for performing network requests.
  * @param error_reporter the error reported protocol to use for reporting errors.
  * @param start_in_sleep_mode true if sleep mode should initialize now
+ * @param issue_callback_configuration optional issue callback configuration.
  */
 logger_id capture_create_logger(
     const char *_Nullable path,
@@ -49,7 +50,8 @@ logger_id capture_create_logger(
     const char *model,
     _Nullable id<Network> network,
     _Nullable id<RemoteErrorReporting> error_reporter,
-    bool start_in_sleep_mode
+    bool start_in_sleep_mode,
+    _Nullable id issue_callback_configuration
 );
 
 /*